Moved hasura functions for cli scripts to shared js
